+++ R/tar_destroy.R
@@ -28,6 +28,7 @@
 #' @inheritSection tar_meta Storage access
 #' @inheritSection tar_unversion Deleting cloud target data
 #' @return `NULL` (invisibly).
+#' @inheritParams tar_delete
 #' @inheritParams tar_validate
 #' @param destroy Character of length 1, what to destroy. Choices:
 #'   * `"all"`: entire data store (default: `_targets/`)
@@ -89,11 +90,15 @@ tar_destroy <- function(
     "workspaces",
     "user"
   ),
+  verbose = TRUE,
   ask = NULL,
   script = targets::tar_config_get("script"),
   store = targets::tar_config_get("store")
 ) {
   tar_assert_allow_meta("tar_destroy", store)
+  tar_assert_lgl(verbose)
+  tar_assert_scalar(verbose)
+  tar_assert_none_na(verbose)
   if (!file.exists(store)) {
     return(invisible())
   }
@@ -116,7 +121,8 @@ tar_destroy <- function(
     tar_delete_cloud_objects(
       names = meta$name,
       meta = meta,
-      path_store = store
+      path_store = store,
+      verbose = verbose
     )
     tar_delete_cloud_meta(script = script)
     unlink(path_scratch_dir_network(), recursive = TRUE)